Allow the LDAP group search to be configured
The IPA LDAP server used in the base cluster only puts full DNs for
users in the member field of group LDAP entries. Allow customization
of the group search template so that the configuration can tell
Gafaelfawr to construct the user DN from a template in order to find
the groups they belong to.