Rework the approach to group membership by DN
Rather than allowing the user to specify a full template for the
group search, instead add a boolean flag saying whether to search
by DNs. This requires that the attributes for retrieving user
information also be set, but I'm hoping to drop support for getting
only groups from LDAP and not also users, so this isn't a major
drawback and it allows simplification of the configuration and
avoids supporting too many degrees of freedom.

Add a test for finding user group membership by DN. Refactor the
LDAP storage layer a little to avoid some duplicate code.

Restructure the validation of the LDAP settings object to take more
advantage of Pydantic v2 validators by using model validators instead
of field validators where more appropriate. Remove all of the defaults
from the frozen config, since that should be determined entirely by
the settings models.