Add new route for CADC token metadata
CADC software uses an authentication component designed for use with
OpenID Connect that takes a token from the Authorization header and
sends it to an endpoint, expecting to get back the validated JWT
claims. Since it treats the token as opaque, we can use this with
our regular tokens as long as we provide an endpoint that returns
user metadata in the expected format.

Add /auth/cadc/userinfo as that endpoint, temporarily. Put CADC in
the URL because the return format is specific to the current needs
of the CADC auth code. In particular, it requires sub to be a UUID
(to allow username changes), which we generate based on an optional
configuration parameter and the UID of the user using the UUID v5
algorithm.