Update Python dependencies
The approach of sharing a database session across all tests and
truncating instead of dropping and recreating the schema between
tests broke horribly and confusingly with the latest version of
pytest-asyncio. Since this was only a minor performance improvement
that was usually dwarfed by other test delays, drop this rather than
fix it, which simplifies the test setup.