Adjust the Redis connection pool parameters
Gafaelfawr is struggling to recover after a Redis server restart,
such as from Kubernetes cluster upgrades. Enable retries on timeout,
which is apparently not enabled by default. Increase the connection
pool size since we've seen one case of pool exhaustion in production.
Increase the length of time Gafaelfawr is willing to wait for a
pool connection to become available.

Use from_pool to construct the Redis object so that we don't have
to separately close the connection pool.