Fix protocol issues in OpenID Connect server
Stop exposing all known claims by default in the issued ID token
and instead honor scopes requested by the client. Currently, the
supported scopes include the profile and email scopes defined by
OpenID Connect (insofar as Gafaelfawr has the data).

Set the aud claim on ID tokens to the client ID rather than a fixed
audience value that matches the Gafaelfawr issuer, bringing the
implementation in line with the intent of the specification.

Tie the expiration time of OpenID Connect ID tokens to the expiration
of the underlying Gafaelfawr token used as an authentication basis.

Require the oidcServer.issuer configuration setting use the https
scheme, since the protocol requires that. Include the scope that was
used for ID token issuance in the response from the token endpoint,
since the spec may require that if unknown scopes were requested.

Declare, in the OpenID configuration endpoint, that the only
supported response mode is query.