Add support for data rights information
Add a rubin scope for the OpenID Connect server and a data_rights
claim that it enables. This claim will contain a space-separated
list of data releases to which the user has access rights, and is
intended for use by IDACs (see DMTN-253).

Add a new config.oidcServer.dataRightsMapping configuration option
that tells Gafaelfawr how to determine those data rights from the
user's group memberships.