Add OpenID Connect return URL registration
In order to allow OpenID Connect clients to be at any domain, not
just one that matches Gafaelfawr's domain, require OpenID Connect
clients to pre-register their return URL. This changes the format
of the oidc-server-secrets secret, since this is now a mandatory
parameter. Verify that it matches the registered client information
in the OIDC routes.