Add support for database migrations with Alembic
Add a new oidc token type as an initial migration, which is not yet
used but will be used in a subsequent change.

Use an async engine for Alembic, since that avoids needing to also
install psycopg2 alongside asyncpg only for Alembic to use it. Since
Alembic only sort-of supports async via asyncio.run, this requires
moving the body of gafaelfawr init into a separate coroutine and
running that with asyncio.run so that Alembic can be run outside of
the running async loop.

The Alembic environment is a bit complicated at the moment and
duplicates a bit of Safir code. A lot of this can be lifted into
Safir in the future.

Fix a few problems with secret loading uncovered by loading the
example configuration: the example OIDC client secrets were missing
the new return_uri field, and newlines weren't removed from the end
of secrets.