Issue oidc tokens for OIDC access tokens
Instead of setting the access token for OpenID Connect authentications
to the same as the ID token, issue a new Gafaelfawr token of the
oidc type with no scopes. Change the userinfo endpoint to expect that
token instead.

This means that the access token will be automatically revoked if
the authentication token used to authenticate the OpenID Connect
login is revoked.