Do not run database initialization on startup
It's not generally safe to run Alembic upgrades on startup since
we may have multiple replicas, and it's not safe to run
gafaelfawr init before Alembic because it will stap the database
with the current version even though it didn't make all the required
changes.

Drop both from the startup script for now. We will need to find a
better way of doing both database initialization and schema changes,
probably using Helm hooks.