Verify the database schema on startup
Unless it is explicitly disabled (such as by the test suite), check
the schema with Alembic during startup to see if it's current and
refuse to start if it isn't. Test this by keeping schema validation
enabled during Selenium testing, which runs via uvicorn and thus
simulates how Gafaelfawr is run in production.

This required changing the Selenium fixtures to not be async, since
Alembic doesn't like async for some things. They turned out to have
no reason to be async, so this was painless.

At present, this only handles web application startup, not the
Kubernetes operator or the audit or maintenance cron jobs.