Improve schema handling and testing
Check whether the schema is current at all entry points, not just
the FastAPI application startup, so that the other aspects of
Gafaelfawr will also not start with an outdated database.

Fix the existing schema migration to use valid SQL. Add proper
tests for schema updates, confirming the ability to upgrade from
the last pre-Alembic schema version.

Add new gafaelfawr update-schema command to initialize the database
if needed and then apply any missing Alembic migrations. Add new
gafaelfawr validate-schema command that exits non-zero if the
database has not been initialized or if the schema is not up to
date.