Add a health check endpoint
Add a new internal endpoint, /health, which checks the database,
Redis, and (if configured) LDAP and raises an uncaught exception if
any of them fail. This will be used as a Kubernetes liveness check
to address the issues we've had with Gafaelfawr recovering from
Redis or LDAP outages and losing its database connections.