Use user info service for health check
The health check tried to bypass the user info service, but that
would require recreating all of the logic that the user info service
already had and bypassed checking connectivity to Firestore.
Instead, add uncached flags to the user info service and its
underlying LDAP and Firestore services that make uncached requests
for a proper health check, and use it.