Drop support for a local development instance
Drop support for starting Gafaelafwr on localhost for development
and replace the documentation for this to instead suggest a Phalanx
development environment.

This support isn't used during normal development and has some
maintenance cost. Deploying Gafaelfawr in a full Phalanx environment
is a better way to do comprehensive integration testing.