Close async database sessions in the operator
Explicitly close async database sessions in the health check and
token portions of the Kubernetes operator. (The ingress handler
doesn't use the database.) Even with an async_scoped_session, the
database session has to be explicitly removed or it will leak and
eventually force the operator to crash and restart.