Remove tokens from Redis if SQL write fails
All token creation paths already created the token in Redis first
and then added an entry to SQL, but if the SQL write failed for
whatever reason, the Gafaelfawr call would fail but the token would
be left behind in Redis. Instead, run the SQL code in a try/except
block and try to remove the newly-created Redis key if the SQL write
failed. This will hopefully reduce the number of orphaned tokens
created during database failures.