Undo the *String variants of Pydantic types
Unfortunately, the code to convert Pydantic Url objects to strings
after validation causes the validation constraints to be lost, so
none of the validation rules are applied to the input URL. Abandon
this as an unworkable idea given how Pydantic currently works and
go back to using regular Url instances for URLs and DSNs in the
configuration.

Add a helper function for creating OIDC client configurations and
use it in the tests instead of manually creating OIDCClient objects,
which is a bit tedious.