Set sub to username for CADC reply
The new CADC authentication code no longer requires the sub claim
returned by the userinfo endpoint to be a UUID, so send the username
instead so that the TAP server will store the username in the UWS
jobs table. This is still not ideal, since the username is not
necessarily a permanent, unique identifier given the possibility of
username changes, but it will do for now.