Cap Gafaelfawr Kopf workers at five
Follow recommendations from a Reddit comment linked in the Kopf
documentation to cap the worker limit at five. We do want to create
Ingress objects in a timely fashion because we use them for labs,
but it doesn't need to be immediate and we don't want to overwhelm
the API server.