Add service parameter to the auth endpoint
Add a top-level service configuration attribute to GafaelfawrIngress
and a service parameter to the auth endpoint. Use this instead of
delegate_to for logging. This is currently optional but will become
mandatory in the future.

If internal token delegation is enabled in GafaelfawrIngress, the
value of the service attribute in the delegation configuration must
match the top-level service attribute. The service attribute inside
the delegation configuration will be removed in the future.

If service and delegate_to are set as parameters to the auth endpoint,
they must have the same values.