Convert datetimes to a new Pydantic type
Defining an external validator for datetime types triggers what I
think is a bug in autodoc_pydantic and can't resolve the Safir
function properly. Defining separate validation and serialization
functions is the old pattern in Pydantic, so instead hide both
beyond a new Annotated type, which accomplishes the same thing
without this problem.