Add a test that the database schema hasn't changed
Add a new test that uses `alembic check` to ensure that the database
schema hasn't changed.

To support maintaining the data files used by that test, add a new
`gafaelfawr generate-schema` command that dumps the SQL commands
required to generate the current schema. Document using that command
to update the most recent schema in the test suite after adding a new
schema migration.