Use uv for Docker build, verify hashes
Switch to uv for Docker builds and enable its hash validation mode
that only verifies hashes that are present. This allows dropping
the update-deps-no-hashes Makefile target, since there's no longer
a need to drop all hashes in order to have one dependency that isn't
hashed (such as a Git dependency).