Overhaul requests from ingress-nginx
Previously, ingress-nginx talked to `/auth`, which was also exposed
via the Gafaelfawr ingress as a regular route. This allows users
to create arbitrary delegated tokens for themselves. Upcoming work
will turn that into a security boundary -- some services will only
be accessible to delegated tokens and not the user directly -- so
that behavior needed to change.

Move `/auth` to `/ingress/auth` and `/auth/anonymous` to
`/ingress/anonymous` so that they won't be exported by the Gafaelfawr
ingress. Modify the Gafaelfawr Kubernetes operator to use an internal
URL for the `auth_request` subhandler instead of the exposed base URL.

As part of this change, make `config.baseUrl` in `GafaelfawrIngress`
optional and get the base and internal base URLs from the Gafaelfawr
configuration by default instead. For now, `config.baseUrl` is still
honored, but it will be removed in a future release.

Drop documentation for manually creating an `Ingress` and only
document `GafaelfawrIngress`. We haven't needed to manually manage
ingresses for quite some time.

Require the `X-Original-URL` header be set for the `/ingress` routes.
ingress-nginx always sets this header, and that change simplifies
some code and documentation.