Allow restricting ingresses to services
Add the concept of a service-only ingress, which in addition to any
other authorization requirements also requires the authentication
token be a token issued to a particular service. Currently, this
means an internal token obtained through a delegation request in its
ingress, although in the future this may include other methods of
associating a token with the service. The key point is that this will
reject authentication directly from users, allowing services to be
limited to requests from other services on behalf of users but not
allow access from those same users directly.

Update some of the `GafaelfawrIngress` documentation to urge setting
`config.service`, mark `config.delegate.internal.service` as deprecated,
and otherwise update for recent changes.