Move token cache transaction management
The pattern of doing database transaction management within the
handlers meant that a database transaction was created in the hot
path of ingress auth requests if delegated tokens were requested.
Move the transaction management into the token cache service so
that the transaction is only created if database activity is
required.