Move all session management to service layer
Stop doing database transaction management in the handlers and
instead move all transaction management into the service layer.
This simplifies the test suite considerably and makes the handlers
simpler and cleaner.

This requires moving the addition of the `admin:token` scope to
new authentications by admins into the token service layer, which
is probably where it always belonged.

In a few cases, the code will open more transactions than it did
before, but all of those cases should be slow paths.