Allow internal tokens on the OIDC userinfo route
Allow a client to present an internal token to the
`/auth/openid/userinfo` endpoint. CADC's authenticator finds the
userinfo endpoint via OpenID Connect configuration and presents
whatever token it has to that endpoint, so this allows it to use the
regular userinfo endpoint.

This will hopefully allow us to drop the `/auth/cadc/userinfo`
route in a future release.