Convert datetime and timedelta to Safir types
Convert the Gafaelfawr `Timestamp` data type to be based on the
Safir `UtcDatetime` instead of using the `normalize_datetime`
function, which will hopefully be deprecated in the future.
Convert some other `datetime` model fields to `UtcDatetime`.

Convert one remaining `timedelta` query parameter to use
`SecondsTimedelta` so that Pydantic will do the validation.