Allow an empty list of required scopes
Allow an authenticated `GafaelfawrIngress` to have an empty list of
required scopes. This is useful for the `onlyService` use case when
any scope is permitted but the requesting token must be a delegated
token for one of the listed services. This was already valid in the
CRD, but previously resulted in a 422 error from Gafaelfawr and a 500
error from ingress-nginx.