Avoid commas in generated auth-url annotation
A `GafaelfawrIngress` requesting ticket delegation with multiple
delegated scopes represented that with comma-separated scopes in the
value of the `delegate_scope` parameter. The current ingress-nginx
doesn't allow any escaped characters, so this causes the `Ingress`
resource to be rejected. Instead, repeat the `delegate_scope` query
parameter for each value, matching how `scope` is handled, which avoids
needing commas.