Convert state handling to Pydantic models
The user's state information, stored in a cookie, was previously
implemented as a dataclass with hand-written serialization and
deserialization code. Replace this with a Pydantic model and validators
and let Pydantic do the JSON serialization and deserialization.

This required making `InvalidTokenError` inherit from `ValueError` as
well as its current parent so that Pydantic would accept exceptions
thrown by validation and process them correctly.