Send users with no login state to after logout
If the user returns to the login route without login state and no return
URL is set (which will be the common case), redirect them to the after
logout URL instead of returning a 403 error. Often this means the user
previously authenticated via another tab and is now logged on, but we
have lost the return URL and do not know where to send them. Returning
the error is more confusing and often causes the user to attempt to
reload the error page, which then fails.