Implement API rate limiting
Add enforcement of API rate limits if the user has an API quota for the
service set in the `GafaelfawrIngress`. If the request is rejected by
rate limiting, Gafaelfawr will return a 429 error with the HTTP
`Retry-After` header set to the time at which the rate limit resets.

Rate limiting is implemented using the `limits` Python package, which
unfortunately introduces a second Redis client with a separate pool
(`coredis`). Currently, rate limiting uses a fixed window without any
fancy rolling expiration or burst protection.