Add support for disallowing notebook spawns
Add a flag to the notebook quota indicating whether the user should be
allowed to spawn a notebook, defaulting to true. This may be set to
false for a specific group (or, later, a quota override) to indicate
that user should not be allowed to spawn notebooks.

This field is calculated and returned by Gafaelfawr but is not enforced
by it. It will be read by Nublado (specifically the Nublado controller).