Tweak Redis connection pool settings
Separate the pool size for the ephemeral Redis pool from the persistent
Redis pool and add a separate pool size for the rate limiting pool
(which alas uses a different Redis client library). Add timeout and
pool size settings for the Redis underlying the limits library, but for
now keep the same retry implementation.