Allow regular users to get quota overrides
The longer-term plan is for squareone to warn if any quota overrides
are in place, which will require the ability to retrieve the overrides
for any authenticated user. Change the permissions on `GET` of the
overrides to any authenticated user, keeping the admin permissions on
`PUT` and `DELETE`, and add a test.