Log quota information
Add quota information to successful authorization log messages, if
applicable. Change the order of logging so that creation of internal
and notebook tokens is logged before the successful authorization,
reflecting the true order of events. Add the service to the logging
context.

Add a log message when a request is rejected due to rate limiting.