Clean up handling of uv run
Use `uv run --only-group=tox` in the GitHub Actions instead of an
explicit `uv sync`. This should also reduce the number of
extraneous dependencies installed. Use the same pattern in the
`Makefile` when updating pre-commit.

Add `--no-dev` to the `uv sync` in `Dockerfile` to ensure that we
don't install development dependencies in the runtime image. Add
`--no-deps` when installing the package itself following a
`uv sync`.

Fix the dependency groups for the `typing` tox environment and stop
using `uv run`, which was silly.

Document that activating the virtualenv may make some editors work
better.