Move cookie parameters into configuration
Make the Gafaelfawr configuration responsible for determining what
flags to pass to `fastapi.Response.set_cookie` in preparation for
adding more complex logic for domain cookies. Always force the cookie
to be secure and drop the logic intended to allow local development
instances, which have not been supported for some time.