Add support for domain-scoped cookies
Allow Gafaelfawr to work across subdomains that are all child
domains of the base URL by accepting a new `allowSubdomains`
configuration parameter and, if it is set, setting domain-scoped
cookies instead of host-scoped cookies. Adjust the validation of
return URLs accordingly, permitting return URLs to a subdomain if
subdomains are allowed.