Check the hosts of ingress rules
Unless the `GafaelfawrIngress` is anonymous or disables cookie
authentication, reject hosts in ingress rules that do not match the
host of the Gafaelfawr base URL or, if `config.allowSubdomains` is
set, is a subdomain of that host. Such ingresses will never work with
browser cookies and may produce confusing redirect loops.