Add support for per-user domain authorization
Add a new setting to `GafaelfawrIngress` named `config.userDomain`
that, if set to true, requires that the last component of the
request URL match the authenticated username. This is intended for
use with per-user subdomains, such as Jupyter labs with JavaScript
origin isolation. In this case, as an additional sanity check, also
require that the hostname in the request be a subdomain of Gafaelfawr's
base domain.