Drop seleniumwire test dependency
seleniumwire has been orphaned and is no longer maintained. Instead
of using it to inject an authentication cookie, instead create a
special route injected by the test suite for Selenium testing that
sets an authentication cookie and visit it automatically. This allows
dropping the seleniumwire dependency and a bunch of workarounds that
were required to keep it.

Update dependencies and inject a structlog logger into the memo for
Kopf runners so that the token operator can log properly and avoid a
Ruff diagnostic.