DM-50591: Fix docker uv cache path
@jonathansick actually [caught this](https://github.com/lsst-sqre/noteburst/pull/116#discussion_r2067200903) when I blindly copied and pasted it into noteburst.

I don't think this will actually change anything about the behavior of the build, because I don't _think_ the `uv pip install --no-deps --compile-bytecode .` command reuses anything from the `uv sync` command, because of the `--no-deps` arg to the `pip install` command. But those paths are probably supposed to match in spirit.