Fix group regex to allow all valid usernames
Fix the group name validation regex so that it accepts all valid
usernames, specifically including usernames that start with
numbers. Otherwise, if `addUserGroup` is enabled, a user with a
username starting with digits would trigger internal validation
errors. If that user's token was the most recent created, this could
in turn cause health check failures.