Randomly choose a user for health checks
Rather than asking the database for one token for health checks,
ask the database for 100 session tokens and randomly choose one
of them. This, coupled with Kubernetes failure tolerance for the
liveness check, should hopefully prevent an isolated problem with a
single user from causing `/health` to reliably fail and bring down
the service.