Base API rate limits on requests per minute
Practical experience with setting rate limits has found that it is
sufficiently undesirable to block a user for up to 15 minutes that
we don't set rate limits for fear of creating that outcome. Given
that and the easier mental math involved in using a more regular
interval, reduce the API rate limit interval to one minute and
interpret API rate limits as requests per minute. One implication is
that the rate limit (apart from total administrative blocks) cannot
be less than one request per minute.