Add TAP quotas
Add a new TAP quota element to user quotas, with the same group math
and override support as API quotas. Currently, there is only one
quota, controlling the allowed number of concurrent queries, but the
data structure is extensible so that other quota information (result
size, perhaps) can be added in the future.