Add new route to get user info without a token
To support TAP quotas in the Qserv Kafka bridge, the bridge needs
to be able to retrieve the quota for a user, but it does not have
a delegated token for the user and we would prefer not to have to
manage that token in the TAP server or send it to Kafka.

Instead, add a new route, `/auth/api/v1/users/<username>`, which
can be used to retrieve the user information for an arbitrary user.
This will also help with administrator debugging of Science Platform
and user authorization issues. Add a new reserved scope,
`admin:userinfo`, that controls access to that endpoint so that
the Qserv Kafka bridge can query it without having full
administrative access.

This route has several limitations: It is not available at all if
Gafaelfawr uses GitHub for authentication (it returns 404), and it
will only show useful information for users in LDAP, even if they
have assigned UIDs or GIDs in Firestore despite not being in LDAP
(as is the case with bot users).