Make default section optional in quota overrides
When setting a quota override, the `default` section can now be
omitted entirely, instead of requiring it be present but empty.
This simplifies overriding the quota for only one group.

Expand the documentation for quotas to add a more detailed example
of how overrides work and to suggest checking the quota for a
specific user after setting an override.