Diagnose and fix malformed Redis token entries
Previously, malformed token entries in Redis would report an
exception to Slack and then act as if the entry didn't exist. This
is correct behavior during authentication, but during audit
Gafaelfawr should detect those entries and report them, and delete
them if the `--fix` flag was provided. Add the correct plumbing to
do that, and refactor the audit method to reduce its complexity.

The audit code is still kind of iffy and needs a more comprehensive
refactor with a better underlying data type.